// src/types/Month.ts
export enum MonthName {
  None = '',
  January = 'January',
  February = 'February',
  March = 'March',
  April = 'April',
  May = 'May',
  June = 'June',
  July = 'July',
  August = 'August',
  September = 'September',
  October = 'October',
  November = 'November',
  December = 'December',
};

export interface MonthType {
  id: string;
  name: MonthName;
};